requirements.txt
streamlit>=1.36.0
pytesseract>=0.3.10
Pillow>=10.3.0
pdf2image>=1.17.0
python-dateutil>=2.9.0.post0
pandas>=2.2.2
regex>=2024.5.15
pyyaml>=6.0
gradio>=4.42.0
PyMuPDF>=1.26.4
